+++ internal/k8s_watcher/watcher.go
@@ -7,14 +7,12 @@ import (
 	"sync"
 	"time"
 
-	"github.com/bitcoin-sv/arc/internal/callbacker/callbacker_api"
 	"github.com/bitcoin-sv/arc/internal/metamorph/metamorph_api"
 )
 
 const (
-	metamorphService  = "metamorph"
-	callbackerService = "callbacker"
-	intervalDefault   = 30 * time.Second
+	metamorphService = "metamorph"
+	intervalDefault  = 30 * time.Second
 )
 
 type K8sClient interface {
@@ -22,14 +20,13 @@ type K8sClient interface {
 }
 
 type Watcher struct {
-	metamorphClient  metamorph_api.MetaMorphAPIClient
-	callbackerClient callbacker_api.CallbackerAPIClient
-	k8sClient        K8sClient
-	logger           *slog.Logger
-	updateInterval   time.Duration
-	namespace        string
-	waitGroup        *sync.WaitGroup
-	cancellations    []context.CancelFunc
+	metamorphClient metamorph_api.MetaMorphAPIClient
+	k8sClient       K8sClient
+	logger          *slog.Logger
+	updateInterval  time.Duration
+	namespace       string
+	waitGroup       *sync.WaitGroup
+	cancellations   []context.CancelFunc
 }
 
 func WithUpdateInterval(d time.Duration) func(*Watcher) {
@@ -41,11 +38,10 @@ func WithUpdateInterval(d time.Duration) func(*Watcher) {
 type ServerOption func(f *Watcher)
 
 // New The K8s watcher listens to events coming from Kubernetes. If it detects a metamorph pod which was terminated, then it sets records locked by this pod to unlocked. This is a safety measure for the case that metamorph is terminated ungracefully where it misses to unlock its records itself.
-func New(logger *slog.Logger, metamorphClient metamorph_api.MetaMorphAPIClient, callbackerClient callbacker_api.CallbackerAPIClient, k8sClient K8sClient, namespace string, opts ...ServerOption) *Watcher {
+func New(logger *slog.Logger, metamorphClient metamorph_api.MetaMorphAPIClient, k8sClient K8sClient, namespace string, opts ...ServerOption) *Watcher {
 	watcher := &Watcher{
-		metamorphClient:  metamorphClient,
-		callbackerClient: callbackerClient,
-		k8sClient:        k8sClient,
+		metamorphClient: metamorphClient,
+		k8sClient:       k8sClient,
 
 		namespace:      namespace,
 		logger:         logger,